Refactor canBookmark function location so data folder is cleaner
